Huang Yihao_832301322_contact_backend/src/requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-CORS==3.0.10
Flask-Migrate==4.0.4
pymysql==1.0.2
cryptography==39.0.1
gunicorn==20.1.0
Werkzeug==2.2.3